refactor (constants and package structure) 1. generate and apply new constants for intent keys 2. generate and apply new constants for saved instance state keys 3. package structure - new ui, data ->model
