+++ app/src/main/AndroidManifest.xml
@@ -10,10 +10,10 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Todolist">
         <activity
-            android:name="com.survivalcoding.todolist.write.SimpleTodoWriteActivity"
+            android:name="com.survivalcoding.todolist.ui.write.SimpleTodoWriteActivity"
             android:exported="false" />
         <activity
-            android:name="com.survivalcoding.todolist.main.MainActivity"
+            android:name="com.survivalcoding.todolist.ui.main.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
