+++ app/src/main/java/com/survivalcoding/todolist/model/TodoItem.kt
@@ -1,4 +1,4 @@
-package com.survivalcoding.todolist.data
+package com.survivalcoding.todolist.model
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
