+++ app/src/main/java/com/survivalcoding/todolist/ui/main/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.survivalcoding.todolist.main
+package com.survivalcoding.todolist.ui.main
 
 import android.app.Activity
 import android.content.Intent
@@ -6,18 +6,23 @@ import android.os.Bundle
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.appcompat.app.AppCompatActivity
 import androidx.recyclerview.widget.LinearLayoutManager
-import com.survivalcoding.todolist.data.TodoItem
+import com.survivalcoding.todolist.model.TodoItem
 import com.survivalcoding.todolist.databinding.ActivityMainBinding
-import com.survivalcoding.todolist.write.SimpleTodoWriteActivity
+import com.survivalcoding.todolist.ui.write.SimpleTodoWriteActivity
+import com.survivalcoding.todolist.ui.write.SimpleTodoWriteActivity.Companion.NEW_TODO
 
 class MainActivity : AppCompatActivity() {
+    companion object {
+        private const val SAVED_TODOS = "todos"
+        const val NEW_ID = "newID"
+    }
     private val binding by lazy {
         ActivityMainBinding.inflate(layoutInflater)
     }
     private val getResult =
         registerForActivityResult(ActivityResultContracts.StartActivityForResult()) {
             if (it.resultCode == Activity.RESULT_OK) {
-                it.data?.getParcelableExtra<TodoItem>("todo")
+                it.data?.getParcelableExtra<TodoItem>(NEW_TODO)
                     ?.let { content -> data.add(content) }
             }
         }
@@ -34,7 +39,7 @@ class MainActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         setContentView(binding.root)
 
-        savedInstanceState?.getParcelableArrayList<TodoItem>("todos")?.let {
+        savedInstanceState?.getParcelableArrayList<TodoItem>(SAVED_TODOS)?.let {
             data.removeAll {true}
             data.addAll(it)
         }
@@ -51,13 +56,13 @@ class MainActivity : AppCompatActivity() {
 
         binding.fab.setOnClickListener {
             val intent = Intent(this, SimpleTodoWriteActivity::class.java)
-            intent.putExtra("newID", data.size)
+            intent.putExtra(NEW_ID, data.size)
             getResult.launch(intent)
         }
     }
 
     override fun onSaveInstanceState(outState: Bundle) {
         super.onSaveInstanceState(outState)
-        outState.putParcelableArrayList("todos", ArrayList(data))
+        outState.putParcelableArrayList(SAVED_TODOS, ArrayList(data))
     }
 }
