+++ app/src/main/java/com/survivalcoding/todolist/ui/write/SimpleTodoWriteActivity.kt
@@ -1,14 +1,18 @@
-package com.survivalcoding.todolist.write
+package com.survivalcoding.todolist.ui.write
 
 import android.app.Activity
 import android.content.Intent
 import android.os.Bundle
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
-import com.survivalcoding.todolist.data.TodoItem
+import com.survivalcoding.todolist.model.TodoItem
 import com.survivalcoding.todolist.databinding.ActivitySimpleTodoWriteBinding
+import com.survivalcoding.todolist.ui.main.MainActivity.Companion.NEW_ID
 
 class SimpleTodoWriteActivity : AppCompatActivity() {
+    companion object {
+        const val NEW_TODO = "newTodo"
+    }
     private val binding by lazy {
         ActivitySimpleTodoWriteBinding.inflate(layoutInflater)
     }
@@ -16,7 +20,7 @@ class SimpleTodoWriteActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(binding.root)
-        val newID = intent.getIntExtra("newID", 0).toLong()
+        val newID = intent.getIntExtra(NEW_ID, 0).toLong()
 
         binding.button.setOnClickListener {
             val title = binding.editTextTitle.text.toString().trim()
@@ -26,7 +30,7 @@ class SimpleTodoWriteActivity : AppCompatActivity() {
                 Toast.makeText(this, "type todo content", Toast.LENGTH_SHORT).show()
             } else {
                 intent = Intent()
-                intent.putExtra("todo", TodoItem(id = newID, title = title, description = description))
+                intent.putExtra(NEW_TODO, TodoItem(id = newID, title = title, description = description))
                 setResult(Activity.RESULT_OK, intent)
                 finish()
             }
