+++ app/src/main/res/layout/activity_simple_todo_write.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".write.SimpleTodoWriteActivity">
+    tools:context=".ui.write.SimpleTodoWriteActivity">
 
     <EditText
         android:id="@+id/editTextTitle"
